+++ utils.hpp
@@ -3,7 +3,7 @@
 
 #include <string>
 
-std::string insertingNumber(std::string r, int number);
+std::string insertingNumber(std::string r, unsigned long long number);
 void printCmd(std::string commandBlock, std::string outputFile);
 
 #endif
